Add email remember feature to login function
Enhance the login functionality by introducing an email remember feature. - Add a new `handleLogin` function that checks if the 'remember me' option is selected by the user. - If selected, the user's email is saved to `localStorage`; otherwise, it is removed. - Update the login process to use `handleLogin` in both mobile and desktop login flows.

Release-Note: Support for remembering user emails added on login.

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>